Use AWS credentials from env

+++ request-ssl.sh
@@ -21,13 +21,14 @@ DOMAIN=$1
 EMAIL=$2
 S3_STORE=$3
 
-
-ROLE=`curl -s http://169.254.169.254/latest/meta-data/iam/security-credentials/`
-curl -s "http://169.254.169.254/latest/meta-data/iam/security-credentials/$ROLE" > /tmp/aws.keys
-export AWS_ACCESS_KEY_ID=`cat /tmp/aws.keys | jq -j '.AccessKeyId'`
-export AWS_SECRET_ACCESS_KEY=`cat /tmp/aws.keys | jq -j '.SecretAccessKey'`
-export AWS_SESSION_TOKEN=`cat /tmp/aws.keys | jq -j '.Token'`
-rm /tmp/aws.keys
+if [ -z $AWS_ACCESS_KEY_ID ] || [ -z $AWS_SECRET_ACCESS_KEY ]; then
+    ROLE=`curl -s http://169.254.169.254/latest/meta-data/iam/security-credentials/`
+    curl -s "http://169.254.169.254/latest/meta-data/iam/security-credentials/$ROLE" > /tmp/aws.keys
+    export AWS_ACCESS_KEY_ID=`cat /tmp/aws.keys | jq -j '.AccessKeyId'`
+    export AWS_SECRET_ACCESS_KEY=`cat /tmp/aws.keys | jq -j '.SecretAccessKey'`
+    export AWS_SESSION_TOKEN=`cat /tmp/aws.keys | jq -j '.Token'`
+    rm /tmp/aws.keys
+fi
 
 if [ ! -d "$HOME/.acme.sh" ]; then
   # curl -s https://get.acme.sh | sh
@@ -35,7 +36,7 @@ if [ ! -d "$HOME/.acme.sh" ]; then
   git clone https://github.com/AntonTimiskov/acme.sh ~/.acme.sh
 fi
 
-echo "ACCOUNT_EMAIL=$EMAIL\n" >> ~/.acme.sh/account.conf
+echo "ACCOUNT_EMAIL=$EMAIL\n" > ~/.acme.sh/account.conf
 
 aws s3 sync $S3_STORE/$DOMAIN/ ~/.acme.sh/$DOMAIN/
